Patch-ioner/src/alphaclip/requirements.txt
# Core dependencies for AlphaCLIP standalone
torch>=1.7.1
torchvision
ftfy
regex
tqdm
loralib
Pillow
numpy
packaging
